move to weldx.constants
