+++ setup.py
@@ -1,7 +1,3 @@
-# Licensed under a 3-clause BSD style license - see LICENSE
-# -*- coding: utf-8 -*-
-"""Python install setup."""
-
 from setuptools import setup
 from setuptools_scm import get_version
