+++ weldx/tests/asdf_tests/test_asdf_core.py
@@ -13,9 +13,8 @@ from fs.osfs import OSFS
 from scipy.spatial.transform import Rotation
 
 import weldx.transformations as tf
-from weldx.asdf.types import META_ATTR
 from weldx.asdf.util import write_buffer, write_read_buffer
-from weldx.constants import Q_
+from weldx.constants import META_ATTR, Q_
 from weldx.core import MathematicalExpression as ME  # nopep8
 from weldx.core import TimeSeries
 from weldx.geometry import SpatialData
